+++ tests/testthat/test_1_5_topicsTest.R
@@ -31,14 +31,71 @@ test_that("topicsTest performs linear regression correctly", {
     save_dir = save_dir_temp
     )
 #  test = result
-  testthat::expect_true(is.list(result[[1]]))
-  testthat::expect_equal(result[[1]]$test_method, "linear_regression")
-  testthat::expect_true(any(grepl("estimate", names(result[[1]]$test))))
-  testthat::expect_true(any(grepl("t", names(result[[1]]$test))))
-  testthat::expect_true(any(grepl("p", names(result[[1]]$test))))
-  testthat::expect_true(any(grepl("p_adjusted", names(result[[1]]$test))))
-    
+  testthat::expect_true(is.list(result))
+  testthat::expect_equal(result$test_method, "linear_regression")
+  testthat::expect_true(any(grepl("estimate", names(result$test))))
+  testthat::expect_true(any(grepl("t", names(result$test))))
+  testthat::expect_true(any(grepl("p", names(result$test))))
+  testthat::expect_true(any(grepl("p_adjusted", names(result$test))))
+  
+  
+  testthat::expect_equal(result$test_method, "linear_regression")
+  
+  testthat::expect_equal(result$test$x.Age.estimate[1:5], 
+           c(-0.02968493, -0.02739334,  0.03571515, -0.01120996, -0.08525859), tolerance = 0.0001)
+
+  testthat::expect_equal(result$test$x.Age.t[1:5], 
+                         c(-0.6627383, -0.6115369,  0.7975251, -0.2501762, -1.9095765), tolerance = 0.0001)
+  
+  testthat::expect_equal(result$test$x.Age.p[1:5], 
+                         c(0.50780478, 0.54112328, 0.42552623, 0.80255430, 0.05676206), tolerance = 0.0001)
+  
+  testthat::expect_equal(result$test$x.Age.p_adjusted[1:5], 
+                         c(0.7459746, 0.7459746, 0.7459746, 0.8917270, 0.3784137), tolerance = 0.0001)
+  
+  testthat::expect_equal(result$test$y.PHQ9tot.estimate[1:5], 
+                         c(-0.028913708,  0.063841350, -0.012663082, -0.048725894,  0.004654584), tolerance = 0.0001)
 })
+
+test_that("topicsTest performs logistic regression correctly", {
+  
+  testthat::skip_on_cran()
+  save_dir_temp <- tempfile()
+  
+  #  data <- dep_wor_data %>% dplyr::mutate(Gender = ifelse(Gender == "male", 1, 0))
+  
+  dtm <- topics::topicsDtm(
+    data = dep_wor_data$Deptext, 
+    save_dir = save_dir_temp) 
+  
+  model <- topics::topicsModel(
+    dtm = dtm, 
+    save_dir = save_dir_temp)
+  
+  preds <- topics::topicsPreds(
+    model = model, 
+    data = dep_wor_data$Deptext, 
+    save_dir = save_dir_temp)
+  
+  result <- topics::topicsTest(
+    model = model, 
+    preds = preds, 
+    data = dep_wor_data, 
+    pred_var_x = "Age", 
+    test_method = "logistic_regression", 
+    save_dir = save_dir_temp)
+  
+  testthat::expect_true(is.list(result))
+  testthat::expect_equal(result$test_method, "logistic_regression")
+  testthat::expect_true(any(grepl("estimate", names(result$test))))
+  testthat::expect_true(any(grepl(".t", names(result$test))))
+  testthat::expect_true(any(grepl(".p", names(result$test))))
+  
+})
+
+#### Test t-test here! 
+
+
 
 
 #test_that("topicsTest handles missing pred_var for non t-test methods", {
@@ -97,10 +154,10 @@ test_that("topicsTest adjusts p-values for multiple comparisons", {
     save_dir = save_dir_temp)
   
   testthat::expect_true(is.list(result))
-  testthat::expect_equal(result[[1]]$test_method, "linear_regression")
-  testthat::expect_true(any(grepl("Age.estimate", names(result[[1]]$test))))
-  testthat::expect_true(any(grepl("Age.t", names(result[[1]]$test))))
-  testthat::expect_true(any(grepl("Age.p", names(result[[1]]$test))))
+  testthat::expect_equal(result$test_method, "linear_regression")
+  testthat::expect_true(any(grepl("Age.estimate", names(result$test))))
+  testthat::expect_true(any(grepl("Age.t", names(result$test))))
+  testthat::expect_true(any(grepl("Age.p", names(result$test))))
   
 })
 
@@ -129,7 +186,7 @@ test_that("topicsTest saves test results to the specified directory", {
     pred_var_x = "Age", 
     save_dir = save_dir_temp)
   
-  testthat::expect_true(file.exists(file.path(save_dir_temp, "seed_42", paste0("test_", result[[1]]$test_method, "_Age.rds"))))
+  testthat::expect_true(file.exists(file.path(save_dir_temp, "seed_42", paste0("test_", result$test_method, "_Age.rds"))))
   
   })
 
@@ -150,49 +207,15 @@ test_that("topicsTest saves test results to the specified directory", {
 #  result <- topicsTest(load_dir = "./results")
 #  result
 #  
-#  testthat::expect_true(is.list(result[[1]]))
-#  testthat::expect_equal(result[[1]]$test_methodtopicsTest performs logistic regression correctly, "linear_regression")
-#  testthat::expect_true("hilstotal.estimate" %in% names(result[[1]]$test))
-#  testthat::expect_true("hilstotal.t" %in% names(result[[1]]$test))
-#  testthat::expect_true("hilstotal.p" %in% names(result[[1]]$test))
+#  testthat::expect_true(is.list(result))
+#  testthat::expect_equal(result$test_methodtopicsTest performs logistic regression correctly, "linear_regression")
+#  testthat::expect_true("hilstotal.estimate" %in% names(result$test))
+#  testthat::expect_true("hilstotal.t" %in% names(result$test))
+#  testthat::expect_true("hilstotal.p" %in% names(result$test))
 #})
 
 
-test_that("topicsTest performs logistic regression correctly", {
-  
-  testthat::skip_on_cran()
-  save_dir_temp <- tempfile()
-  
-#  data <- dep_wor_data %>% dplyr::mutate(Gender = ifelse(Gender == "male", 1, 0))
-  
-  dtm <- topics::topicsDtm(
-    data = dep_wor_data$Deptext, 
-    save_dir = save_dir_temp) 
-  
-  model <- topics::topicsModel(
-    dtm = dtm, 
-    save_dir = save_dir_temp)
-  
-  preds <- topics::topicsPreds(
-    model = model, 
-    data = dep_wor_data$Deptext, 
-    save_dir = save_dir_temp)
-  
-  result <- topics::topicsTest(
-    model = model, 
-    preds = preds, 
-    data = dep_wor_data, 
-    pred_var_x = "Age", 
-    test_method = "logistic_regression", 
-    save_dir = save_dir_temp)
-  
-  testthat::expect_true(is.list(result[[1]]))
-  testthat::expect_equal(result[[1]]$test_method, "logistic_regression")
-  testthat::expect_true(any(grepl("estimate", names(result[[1]]$test))))
-  testthat::expect_true(any(grepl(".t", names(result[[1]]$test))))
-  testthat::expect_true(any(grepl(".p", names(result[[1]]$test))))
- 
-})
+
 
 #test_that("topicsTest performs ridge regression correctly", {
 #  testthat::skip_on_cran()
@@ -201,10 +224,10 @@ test_that("topicsTest performs logistic regression correctly", {
 #  preds <- topicsPreds(model = model, data = data$Deptext)
 #  
 #  result <- topicsTest(model = model, preds = preds, data = data, pred_var_x = "Age", test_method = "ridge_regression")
-#  result[[1]]
-#  expect_true(is.list(result[[1]]))
-#  expect_equal(result[[1]]$test_method, "ridge_regression")
-#  expect_true(any(grepl("estimate", names(result[[1]]$test))))
-#  expect_true(any(grepl("statistic", names(result[[1]]$test))))
-#  expect_true(any(grepl("p.value", names(result[[1]]$test))))
+#  result
+#  expect_true(is.list(result))
+#  expect_equal(result$test_method, "ridge_regression")
+#  expect_true(any(grepl("estimate", names(result$test))))
+#  expect_true(any(grepl("statistic", names(result$test))))
+#  expect_true(any(grepl("p.value", names(result$test))))
 #})